ZZALOG: Convert QRZ API LogID to unsigned long long from int. Added new callback cb_value_ull to get an unsigned long long from a Fl_Int_Input widget.

+++ zzalog/include/web_dialog.h
@@ -128,7 +128,7 @@ class Fl_Group;
 		struct api_logbook_data {
 			bool used;
 			string key;
-			int last_logid;
+			unsigned long long last_logid;
 			string last_download;
 		};
 		map<string, api_logbook_data*> qrz_api_data_;
